+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Fix crash where trying to close the MQTT connection whilst it's connecting thows an unhandled exception
 - Only latest stop reason should be printed to logs on startup
 - Fix bug where geofencing client wasn't initialized properly, leading to very unreliable region transition detection
+- Fix bug where some settings (`pubQos`, `mqttProtocolLevel` etc.) couldn't be set via the config editor (#1801)
 
 ## Version 2.5.1